Разделяет код
